Update Microsoft.Security.Utilities.Core from v1.17.0 to v1.18.0 (#5224)
Release Notes: https://github.com/microsoft/security-utilities/blob/release/v1.18.0/docs/ReleaseHistory.md

This release includes many new granular secret masking rules and significant performance enhancements. Benchmarks show a net speedup.

It also improves the case where a literal secret value is added to the masker that also matches a rule. In this case, redaction will use `***` instead of the rule-based `SECNNN/NNN:ID`. A new agent test is also added in this change to verify this behavior.

The update required some refactoring to absorb the removal of `SecretMasker.Clone`. The agent secret masker no longer implements the server `ISecretMasker` interface that has a `Clone` method. There was only one place where the agent masker was passed to server `ISecretMasker` interface, but this was deemed to be unnecessary as we also pass an `ITrace` that handles secret masking. We therefore now pass `null` in this case with a detailed comment explaining why this is correct. A new test is added to verify that masking is still performed in this code path. 

There is also some minor refactoring in tests to provide a shared `TestHostContext.GetTraceContent` method. A handful of tests with duplicated code are refactored to use this and a new test takes advantage of it.

Finally, `TestHostContext` is updated to use the non-legacy `OssSecretMasker`.